pyproject.toml
[project]
name = "antler"
version = "0.0.0"
description = "Development utilities for the Antler rendering engine"
authors = [{ name = "Freddy Wordingham", email = "freddy@wordingham.ai" }]
readme = "README.md"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
pillow = ">=12.2,<13.0"
python = ">=3.12,<4.0"
typeguard = ">=4.5,<5.0"
